requirements.txt
streamlit
pandas
joblib
python-dotenv
google-generativeai
scikit-learn
nltk
langchain
chromadb
langchain-community
sentence-transformers
pypdf
pysqlite3-binary
langchain-google-genai
git+https://github.com/coleifer/pysqlite3.git

